+++ tests/testthat/test.devices.R
@@ -2,13 +2,10 @@ context("Devices")
 library(tregata)
 
 data("gasmet", "meta")
-meta1 <- meta
-meta1$pmbar[3] <- NA
 meta$temp[3] <- NA
 
 test_that("Flux calculation stops", {
-  expect_error(preprocess_gasmet(gasmet,meta), "missing preassure or temperature values in meta file. Flux calculation stoped.")
-  expect_error(preprocess_gasmet(gasmet,meta1), "missing preassure or temperature values in meta file. Flux calculation stoped.")
+  expect_error(preprocess_gasmet(gasmet,meta), "missing temperature values in meta file. Flux calculation stoped.")
 })
 
 rm(meta)