+++ NEWS.md
@@ -3,6 +3,8 @@
 * add functions to create ggplot2 ylabs for gasfluxes
 * add preprocessing of GASMET data
 * add complete processing of GASMET data
+* use atm. pressure data from GASMET data directly
+* add routines to read GASMET and LosGatos files from disk
 
 # tregata 0.1.0
 * add basic functions